requirements.txt
numpy==1.19.2
pandas==1.1.2
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
